requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
requests>=2.31.0
feedparser>=6.0.10
jinja2>=3.1.0
